% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reldist.r
\name{reldist}
\alias{reldist}
\title{Calculate the relative distance distribution b/w two feature files.}
\usage{
reldist(a, b, detail = NULL)
}
\arguments{
\item{a}{<bed/gff/vcf>}

\item{b}{<bed/gff/vcf>}

\item{detail}{Instead of a summary, report the relative   distance for each interval in A}
}
\description{
Calculate the relative distance distribution b/w two feature files.
}
